# backend/docs/swagger.yaml
basePath: /api
definitions:
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      onboarding_completed:
        type: boolean
    type: object
...
  handler.recordViewRequest:
    properties:
      item_id:
        type: string
      item_type:
        enum:
        - book
        - movie
        type: string
    required:
    - item_id
    - item_type
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Cross-media recommendation system API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Recly API
  version: "1.0"
paths:
...
  /genres:
    get:
      parameters:
      - default: all
        description: Item type (book/movie/all)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: genres
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get genres
      tags:
      - search
  /items/batch:
    get:
      parameters:
      - description: Comma-separated IDs
        in: query
        name: ids
        required: true
        type: string
      - default: all
        description: Item type (book/movie/all)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get multiple items by IDs
      tags:
      - search
...
  /search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: all
        description: Item type (book/movie/all)
        in: query
        name: type
        type: string
      - description: Genre filter
        in: query
        name: genre
        type: string
      - description: 'Sort field: ratings_count:desc, vote_count:desc'
        in: query
        name: sort
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: results
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Search items
      tags:
      - search
  /user/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar image (jpg, png, webp)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: avatar_url
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Upload avatar
      tags:
      - user
...
  /user/onboarding/complete:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Complete onboarding
      tags:
      - user
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: user profile
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties: true
            type: object
      summary: Get user profile
      tags:
      - user
...
  /user/saved-items:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: saved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get saved items
      tags:
      - saved
    post:
      consumes:
      - application/json
      parameters:
      - description: Item data (item_type, item_id)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Save item
      tags:
      - saved
  /user/saved-items/{id}:
    delete:
      parameters:
      - description: Saved item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Delete saved item
      tags:
      - saved
  /user/view:
    post:
      consumes:
      - application/json
      parameters:
      - description: View data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.recordViewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Record view
      tags:
      - views
  /user/views:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: views
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get recent views
      tags:
      - views
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"